backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg[binary]==3.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.4.0
redis==5.0.8
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

